pyproject.toml
[tool.poetry]
name = "obsidianbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
graia-ariadne = {extras = ["standard"], version = "^0.10.2"}
graia-saya = "^0.0.17"
aiohttp = "^3.8.3"
pyyaml = "^6.0"
requests-cache = "^0.9.7"
pymysql = "^1.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
[[tool.poetry.source]]

name = "tuna-tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = false
